Complete porting of REST tests
